Move out SortedInt32KeyNode.Enumerator
