+++ src/System.Collections.Immutable/src/System.Collections.Immutable.csproj
@@ -34,7 +34,6 @@
     <Compile Include="System\Collections\Immutable\IImmutableStack.cs" />
     <Compile Include="System\Collections\Immutable\IStrongEnumerable_2.cs" />
     <Compile Include="System\Collections\Immutable\IStrongEnumerator_1.cs" />
-    <Compile Include="System\Collections\Immutable\SortedInt32KeyNode.cs" />
     <Compile Include="System\Collections\Immutable\IOrderedCollection.cs" />
     <Compile Include="System\Collections\Immutable\ImmutableArray.cs" />
     <Compile Include="System\Collections\Immutable\ImmutableArray_1.Builder.cs" />
@@ -88,6 +87,8 @@
     <Compile Include="System\Collections\Immutable\KeysOrValuesCollectionAccessor.cs" />
     <Compile Include="System\Collections\Immutable\RefAsValueType.cs" />
     <Compile Include="System\Collections\Immutable\SecureObjectPool.cs" />
+    <Compile Include="System\Collections\Immutable\SortedInt32KeyNode.cs" />
+    <Compile Include="System\Collections\Immutable\SortedInt32KeyNode.Enumerator.cs" />
     <Compile Include="System\Linq\ImmutableArrayExtensions.cs" />
     <Compile Include="Validation\Requires.cs" />
     <Compile Include="Validation\ValidatedNotNullAttribute.cs" />
